Create sketch_multi_fx
Teensy .ino sketch to be used with the Episodes pedal. This sketch combines the Glitch Delay, Ghost Code and Acid Rain sketches. The toggle switch selects between the three.

You need to donwload the additional libraries (saved on Github but more easily accessible through my Wix site) for this sketch to run.